move `types` to `pkg` from `internal`
