+++ internal/python/plan.go
@@ -1,11 +1,12 @@
 package python
 
 import (
-	. "github.com/zeabur/zbpack/internal/types"
-	"github.com/zeabur/zbpack/internal/utils"
 	"os"
 	"path"
 	"strings"
+
+	"github.com/zeabur/zbpack/internal/utils"
+	. "github.com/zeabur/zbpack/pkg/types"
 )
 
 func DetermineFramework(absPath string) PythonFramework {
