+++ internal/python/python.go
@@ -1,8 +1,9 @@
 package python
 
 import (
-	"github.com/zeabur/zbpack/internal/types"
 	"os"
+
+	"github.com/zeabur/zbpack/pkg/types"
 )
 
 func GenerateDockerfile(meta types.PlanMeta) (string, error) {
