+++ internal/static/static.go
@@ -1,6 +1,6 @@
 package static
 
-import "github.com/zeabur/zbpack/internal/types"
+import "github.com/zeabur/zbpack/pkg/types"
 
 func GenerateDockerfile(meta types.PlanMeta) (string, error) {
 	dockerfile := `FROM nginx:alpine
